Moved CustomJWTAuthentication to custom_authentication.py file , changed url_path - (right_to_draw to right_draw)
