+++ pcb_design/authentication/custom_permissions.py
@@ -1,9 +1,5 @@
 from rest_framework import permissions
 from django.contrib.auth.models import Permission
-from rest_framework_simplejwt.authentication import JWTAuthentication
-from rest_framework.exceptions import AuthenticationFailed
-
-
 class IsAuthorized(permissions.BasePermission):
 
     def has_permission(self, request, view):
@@ -37,14 +33,3 @@ class IsAuthorized(permissions.BasePermission):
             return True
         
         return False
-    
-
-class CustomJWTAuthentication(JWTAuthentication):
-    def authenticate(self, request):        
-        user, token = super().authenticate(request)
-
-        # Check if the user is logged out
-        if user and user.is_logged_out:
-            raise AuthenticationFailed("User is logged out")
-
-        return user, token
